LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/display-contents-expected.txt


FAIL getComputedStyle when container is display:contents assert_equals: expected "" but got "50"
FAIL getComputedStyle when container becomes display:contents assert_equals: expected "30" but got "50"
FAIL getComputedStyle when intermediate container becomes display:contents assert_equals: expected "30" but got "50"
